refactor(console): separate the default and Lambda consoles
