+++ llrt_core/Cargo.toml
@@ -13,42 +13,10 @@ macro = ["rquickjs/macro"]
 bindgen = ["rquickjs/bindgen"]
 
 [dependencies]
-llrt_context = { path = "../libs/llrt_context" }
-llrt_encoding = { path = "../libs/llrt_encoding" }
-llrt_json = { path = "../libs/llrt_json" }
-llrt_modules = { path = "../llrt_modules", features = ["all"] }
-llrt_numbers = { path = "../libs/llrt_numbers" }
-llrt_utils = { path = "../libs/llrt_utils", features = ["all"] }
-chrono = { version = "0.4", default-features = false, features = ["std"] }
-quick-xml = "0.37"
-phf = "0.11"
-
-rquickjs = { git = "https://github.com/DelSkayn/rquickjs.git", version = "0.9.0", features = [
-  "full-async",
-  "parallel",
-  "rust-alloc",
-], default-features = false }
-tokio = { version = "1", features = ["full"] }
-tracing = { version = "0.1", features = ["log"] }
-uuid = { version = "1", default-features = false, features = [
-  "v1",
-  "v3",
-  "v4",
-  "v5",
-  "v6",
-  "v7",
-  "fast-rng",
-] }
-once_cell = "1"
-zstd = { version = "0.13", default-features = false, features = [] }
-libc = "0.2"
 bytes = "1"
-uuid-simd = "0.8"
-simd-json = { version = "0.14", default-features = false, features = [
-  "big-int-as-float",
-] }
-itoa = "1"
-ryu = "1"
+chrono = { version = "0.4", default-features = false, features = ["std"] }
+home = "0.5"
+http-body-util = "0.1"
 hyper = { version = "1", features = ["client", "http1", "http2"] }
 hyper-rustls = { version = "0.27", default-features = false, features = [
   "http2",
@@ -58,16 +26,49 @@ hyper-rustls = { version = "0.27", default-features = false, features = [
   "ring",
 ] }
 hyper-util = "0.1"
-http-body-util = "0.1"
+itoa = "1"
+libc = "0.2"
+llrt_context = { path = "../libs/llrt_context" }
+llrt_console = { path = "../modules/llrt_console" }
+llrt_encoding = { path = "../libs/llrt_encoding" }
+llrt_json = { path = "../libs/llrt_json" }
+llrt_logging = { path = "../libs/llrt_logging" }
+llrt_modules = { path = "../llrt_modules", features = ["lambda"] }
+llrt_numbers = { path = "../libs/llrt_numbers" }
+llrt_utils = { path = "../libs/llrt_utils", features = ["all"] }
+once_cell = "1"
+phf = "0.11"
+quick-xml = "0.37"
+ring = "0.17"
+rquickjs = { git = "https://github.com/DelSkayn/rquickjs.git", version = "0.9.0", features = [
+  "full-async",
+  "parallel",
+  "rust-alloc",
+], default-features = false }
 rustls = { version = "0.23", default-features = false, features = [
   "tls12",
   "ring",
 ] }
-ring = "0.17"
 rustls-pemfile = "2"
-home = "0.5"
-url = "=2.5.1"
+ryu = "1"
+simd-json = { version = "0.14", default-features = false, features = [
+  "big-int-as-float",
+] }
 terminal_size = "0.4"
+tokio = { version = "1", features = ["full"] }
+tracing = { version = "0.1", features = ["log"] }
+url = "=2.5.1"
+uuid = { version = "1", default-features = false, features = [
+  "v1",
+  "v3",
+  "v4",
+  "v5",
+  "v6",
+  "v7",
+  "fast-rng",
+] }
+uuid-simd = "0.8"
+zstd = { version = "0.13", default-features = false, features = [] }
 
 [target.'cfg(target_os = "windows")'.dependencies]
 md-5 = { version = "0.10" }
