+++ modules/llrt_console/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "llrt_console"
+description = "LLRT Module console"
+version = "0.5.1-beta"
+edition = "2021"
+license = "Apache-2.0"
+repository = "https://github.com/awslabs/llrt"
+
+[lib]
+name = "llrt_console"
+path = "src/lib.rs"
+
+[dependencies]
+llrt_logging = { version = "0.5.1-beta", path = "../../libs/llrt_logging" }
+llrt_utils = { version = "0.5.1-beta", path = "../../libs/llrt_utils", default-features = false }
+rquickjs = { git = "https://github.com/DelSkayn/rquickjs.git", version = "0.9.0", features = [
+  "macro",
+], default-features = false }
